@@ -2,21 +2,8 @@
   "name": "server",
   "version": "0.0.1",
   "description": "Awesome project developed with TypeORM.",
-  "devDependencies": {
-    "@types/bcryptjs": "^2.4.2",
-    "@types/cookie-parser": "^1.4.2",
-    "@types/cors": "^2.8.10",
-    "@types/express": "^4.17.11",
-    "@types/graphql": "^14.5.0",
-    "@types/graphql-upload": "^8.0.4",
-    "@types/jsonwebtoken": "^8.5.1",
-    "@types/node": "^14.14.35",
-    "nodemon": "^2.0.7",
-    "rimraf": "^3.0.2",
-    "typescript": "4.2.3"
-  },
   "dependencies": {
-    "apollo-server-express": "^2.21.2",
+    "@apollo/server": "^4.10.4",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "class-validator": "^0.13.1",
@@ -26,18 +13,33 @@
     "dotenv": "^8.2.0",
     "dotenv-safe": "^8.2.0",
     "express": "^4.17.1",
-    "graphql": "^15.5.0",
+    "graphql": "^16.0.0",
+    "graphql-scalars": "^1.23.0",
     "jsonwebtoken": "^8.5.1",
     "pg": "^8.4.0",
     "pg-connection-string": "^2.4.0",
     "reflect-metadata": "^0.1.10",
-    "ts-node": "^9.1.1",
+    "ts-node": "^10.0.0",
     "twilio": "^3.61.0",
-    "type-graphql": "^1.1.1",
+    "type-graphql": "^2.0.0-rc.1",
     "typeorm": "^0.2.32"
+  },
+  "devDependencies": {
+    "@types/bcryptjs": "^2.4.2",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/cors": "^2.8.10",
+    "@types/express": "^4.17.11",
+    "@types/graphql": "^14.5.0",
+    "@types/graphql-upload": "^8.0.4",
+    "@types/jsonwebtoken": "^8.5.1",
+    "@types/node": "^20.0.0",
+    "nodemon": "^2.0.7",
+    "rimraf": "^3.0.2",
+    "typescript": "5.5.3"
   },
   "scripts": {
     "start": "NODE_ENV=development nodemon --exec ts-node src/index.ts",
-    "build": "rimraf dist && tsc"
+    "build": "rimraf dist && tsc",
+    "dev-up": "docker-compose up -d"
   }
 }
