@@ -1,8 +1,8 @@
 import 'reflect-metadata';
 import 'dotenv/config';
 import express from 'express';
-
-import { ApolloServer } from 'apollo-server-express';
+import { expressMiddleware } from '@apollo/server/express4';
+import { ApolloServer } from '@apollo/server';
 import { buildSchema } from 'type-graphql';
 import { UserResolver } from './resolvers/UserResolver';
 import { ProfileResolver } from './resolvers/ProfileResolver';
@@ -87,22 +87,19 @@ import { Invitation } from './entity/Invitation';
 
   await createTypeOrmConn();
 
-  const apolloServer = new ApolloServer({
-    schema: await buildSchema({
-      resolvers: [
-        UserResolver,
-        ProfileResolver,
-        SubscriptionResolver,
-        PostResolver,
-        LikeResolver,
-        InvitationResolver
-      ]
-    }),
-    context: ({ req, res }) => ({ req, res }),
-    playground: true
-  });
+  const schema = await buildSchema({resolvers: [
+    UserResolver,
+    ProfileResolver,
+    SubscriptionResolver,
+    PostResolver,
+    LikeResolver,
+    InvitationResolver
+  ]})
 
-  apolloServer.applyMiddleware({ app, cors: false });
+  const server = new ApolloServer({schema});
+  await server.start()
+
+  app.use('/graphql', cors<cors.CorsRequest>(), express.json(), expressMiddleware(server));
 
   app.listen(process.env.PORT || 4000, () => {
     console.log(
@@ -110,21 +107,3 @@ import { Invitation } from './entity/Invitation';
     );
   });
 })();
-
-// createConnection()
-//   .then(async (connection) => {
-//     console.log('Inserting a new user into the database...');
-//     const user = new User();
-//     user.firstName = 'Timber';
-//     user.lastName = 'Saw';
-//     user.age = 25;
-//     await connection.manager.save(user);
-//     console.log('Saved a new user with id: ' + user.id);
-
-//     console.log('Loading users from the database...');
-//     const users = await connection.manager.find(User);
-//     console.log('Loaded users: ', users);
-
-//     console.log('Here you can setup and run express/koa/any other framework.');
-//   })
-//   .catch((error) => console.log(error));
